+++ packages/gatsby-plugin-orion-view/CHANGELOG.md
@@ -3,6 +3,64 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.13.0](https://github.com/nearform/orion/compare/v2.12.0...v2.13.0) (2020-02-20)
+
+
+### Features
+
+* 🎸 add AppBar component ([385cac7](https://github.com/nearform/orion/commit/385cac73367c1d3c7eb2d9813f9462eafef9c536))
+* 🎸 make app bar brand logo image set by theme css ([cada92a](https://github.com/nearform/orion/commit/cada92a9cb7ed1df268caf8049248f9c3e20aea4))
+
+
+
+
+
+# [2.12.0](https://github.com/nearform/orion/compare/v2.11.0...v2.12.0) (2020-02-19)
+
+**Note:** Version bump only for package gatsby-plugin-orion-view
+
+
+
+
+
+# [2.11.0](https://github.com/nearform/orion/compare/v2.10.0...v2.11.0) (2020-02-18)
+
+
+### Features
+
+* 🎸 article content view component ([509533b](https://github.com/nearform/orion/commit/509533bcccae15249ed8b58798eb38383629b223))
+
+
+
+
+
+# [2.10.0](https://github.com/nearform/orion/compare/v2.9.1...v2.10.0) (2020-02-18)
+
+
+### Features
+
+* 🎸 article metadata component ([ab35468](https://github.com/nearform/orion/commit/ab3546806e076fe8c1525f578b87e208d816950b))
+
+
+
+
+
+# [2.9.0](https://github.com/nearform/orion/compare/v2.8.0...v2.9.0) (2020-02-13)
+
+**Note:** Version bump only for package gatsby-plugin-orion-view
+
+
+
+
+
+# [2.8.0](https://github.com/nearform/orion/compare/v2.7.0...v2.8.0) (2020-02-13)
+
+**Note:** Version bump only for package gatsby-plugin-orion-view
+
+
+
+
+
 # [2.7.0](https://github.com/nearform/orion/compare/v2.6.0...v2.7.0) (2020-02-13)
 
 **Note:** Version bump only for package gatsby-plugin-orion-view