+++ packages/knowledge-base/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.13.0](https://github.com/nearform/orion/compare/v2.12.0...v2.13.0) (2020-02-20)
+
+**Note:** Version bump only for package knowledge-base
+
+
+
+
+
+# [2.12.0](https://github.com/nearform/orion/compare/v2.11.0...v2.12.0) (2020-02-19)
+
+**Note:** Version bump only for package knowledge-base
+
+
+
+
+
+# [2.11.0](https://github.com/nearform/orion/compare/v2.10.0...v2.11.0) (2020-02-18)
+
+**Note:** Version bump only for package knowledge-base
+
+
+
+
+
+# [2.10.0](https://github.com/nearform/orion/compare/v2.9.1...v2.10.0) (2020-02-18)
+
+**Note:** Version bump only for package knowledge-base
+
+
+
+
+
+# [2.9.0](https://github.com/nearform/orion/compare/v2.8.0...v2.9.0) (2020-02-13)
+
+**Note:** Version bump only for package knowledge-base
+
+
+
+
+
+# [2.8.0](https://github.com/nearform/orion/compare/v2.7.0...v2.8.0) (2020-02-13)
+
+**Note:** Version bump only for package knowledge-base
+
+
+
+
+
 # [2.7.0](https://github.com/nearform/orion/compare/v2.6.0...v2.7.0) (2020-02-13)
 
 **Note:** Version bump only for package knowledge-base